+++ src/Plugin/FlightOverlay.cs
@@ -43,45 +43,47 @@ namespace Trajectories
 
         private sealed class TargetingCross : MonoBehaviour
         {
-            internal const double markerSize = 2.0d; // in meters
+            internal const double MARKER_SIZE = 2.0d; // in meters
 
-            private static double impactLat = 0d;
-            private static double impactLon = 0d;
-            private static double impactAlt = 0d;
-            private static Vector3 screen_point;
-            private static Vector3 cam_pos;
-            private static double cross_dist = 0d;
+            private double latitude = 0d;
+            private double longitude = 0d;
+            private double altitude = 0d;
+            private Vector3 screen_point;
+            private Vector3 cam_pos;
+            private double cross_dist = 0d;
 
-            internal Vector3? ImpactPosition { get; set; }
-            internal CelestialBody ImpactBody { get; set; }
+            internal Vector3? Position { get; set; }
+            internal CelestialBody Body { get; set; }
+            internal Color Color { get; set; } = Color.red;
 
             internal void OnPostRender()
             {
-                if (ImpactPosition == null || ImpactBody == null)
+                if (Position == null || Body == null)
                     return;
 
                 // get impact position, translate to latitude and longitude
-                ImpactBody.GetLatLonAlt(ImpactPosition.Value + ImpactBody.position, out impactLat, out impactLon, out impactAlt);
+                Body.GetLatLonAlt(Position.Value + Body.position, out latitude, out longitude, out altitude);
 
                 // only draw if visible on the camera
-                screen_point = PlanetariumCamera.Camera.WorldToViewportPoint(ImpactPosition.Value + ImpactBody.position);
+                screen_point = PlanetariumCamera.Camera.WorldToViewportPoint(Position.Value + Body.position);
                 if (!(screen_point.z > 0 && screen_point.x > 0 && screen_point.x < 1 && screen_point.y > 0 && screen_point.y < 1))
                     return;
 
                 // resize marker in respect to distance from camera.
-                cam_pos = ScaledSpace.ScaledToLocalSpace(PlanetariumCamera.Camera.transform.position) - ImpactBody.position;
-                cross_dist = System.Math.Max(Vector3.Distance(cam_pos, ImpactPosition.Value) / 80.0d, 1.0d);
+                cam_pos = ScaledSpace.ScaledToLocalSpace(PlanetariumCamera.Camera.transform.position) - Body.position;
+                cross_dist = System.Math.Max(Vector3.Distance(cam_pos, Position.Value) / 80.0d, 1.0d);
 
                 // draw ground marker at this position
-                GLUtils.DrawGroundMarker(ImpactBody, impactLat, impactLon, Color.red, false, 0, Math.Min(markerSize * cross_dist, 1500.0d));
+                GLUtils.DrawGroundMarker(Body, latitude, longitude, Color, false, 0, Math.Min(MARKER_SIZE * cross_dist, 1500.0d));
             }
         }
 
-        private const int defaultVertexCount = 32;
+        private const int DEFAULT_VERTEX_COUNT = 32;
         //private const float lineWidth = 2.0f;
 
         private static TrajectoryLine line;
-        private static TargetingCross targetingCross;
+        private static TargetingCross impact_cross;
+        private static TargetingCross target_cross;
 
         // update method variables, put here to stop over use of the garbage collector.
         private static double time = 0d;
@@ -95,7 +97,9 @@ namespace Trajectories
         {
             Util.DebugLog("Constructing");
             line = FlightCamera.fetch.mainCamera.gameObject.AddComponent<TrajectoryLine>();
-            targetingCross = FlightCamera.fetch.mainCamera.gameObject.AddComponent<TargetingCross>();
+            impact_cross = FlightCamera.fetch.mainCamera.gameObject.AddComponent<TargetingCross>();
+            target_cross = FlightCamera.fetch.mainCamera.gameObject.AddComponent<TargetingCross>();
+            target_cross.Color = Color.green;
         }
 
         internal static void Destroy()
@@ -104,17 +108,22 @@ namespace Trajectories
             if (line != null)
                 UnityEngine.Object.Destroy(line);
 
-            if (targetingCross != null)
-                UnityEngine.Object.Destroy(targetingCross);
+            if (impact_cross != null)
+                UnityEngine.Object.Destroy(impact_cross);
+
+            if (target_cross != null)
+                UnityEngine.Object.Destroy(target_cross);
 
             line = null;
-            targetingCross = null;
+            impact_cross = null;
+            target_cross = null;
         }
 
         internal static void Update()
         {
             line.enabled = false;
-            targetingCross.enabled = false;
+            impact_cross.enabled = false;
+            target_cross.enabled = false;
 
             if (!Settings.DisplayTrajectories
                 || Util.IsMap
@@ -138,9 +147,9 @@ namespace Trajectories
             else
             {
                 time = lastPatch.StartingState.Time;
-                time_increment = (lastPatch.EndTime - lastPatch.StartingState.Time) / defaultVertexCount;
+                time_increment = (lastPatch.EndTime - lastPatch.StartingState.Time) / DEFAULT_VERTEX_COUNT;
                 orbit = lastPatch.SpaceOrbit;
-                for (uint i = 0; i < defaultVertexCount; ++i)
+                for (uint i = 0; i < DEFAULT_VERTEX_COUNT; ++i)
                 {
                     vertex = Util.SwapYZ(orbit.getRelativePositionAtUT(time));
                     if (Settings.BodyFixedMode)
@@ -157,16 +166,30 @@ namespace Trajectories
             line.Body = lastPatch.StartingState.ReferenceBody;
             line.enabled = true;
 
+            // red impact cross
             if (lastPatch.ImpactPosition != null)
             {
-                targetingCross.ImpactPosition = lastPatch.ImpactPosition.Value;
-                targetingCross.ImpactBody = lastPatch.StartingState.ReferenceBody;
-                targetingCross.enabled = true;
+                impact_cross.Position = lastPatch.ImpactPosition.Value;
+                impact_cross.Body = lastPatch.StartingState.ReferenceBody;
+                impact_cross.enabled = true;
             }
             else
             {
-                targetingCross.ImpactPosition = null;
-                targetingCross.ImpactBody = null;
+                impact_cross.Position = null;
+                impact_cross.Body = null;
+            }
+
+            // green target cross
+            if (TargetProfile.WorldPosition != null)
+            {
+                target_cross.Position = TargetProfile.WorldPosition.Value;
+                target_cross.Body = TargetProfile.Body;
+                target_cross.enabled = true;
+            }
+            else
+            {
+                target_cross.Position = null;
+                target_cross.Body = null;
             }
         }
     }